types: ARRAY_LEN should return a size_t value
Instead of pulling in just the one header to get size_t defined, we pull in
the whole int.h header.  It's simpler, and it gets us closer to types.h becoming
the one header to include to get a sane environment.

+++ include/jeffpc/types.h
@@ -23,6 +23,8 @@
 #ifndef __JEFFPC_TYPES_H
 #define __JEFFPC_TYPES_H
 
-#define ARRAY_LEN(x)		(sizeof(x) / sizeof(x[0]))
+#include <jeffpc/int.h>
+
+#define ARRAY_LEN(x)		((size_t) (sizeof(x) / sizeof(x[0])))
 
 #endif
